Move vector structs to seperate cs file
